Change Notes: - HircItem into Hirc folder. - Moved DidxChunk into its own file.
