+++ Editors/Audio/BnkCompiler/ObjectGeneration/IWwiseHircGenerator.cs
@@ -1,5 +1,5 @@
 ﻿using System;
-using Shared.GameFormats.Wwise;
+using Shared.GameFormats.Wwise.Hirc;
 
 namespace Editors.Audio.BnkCompiler.ObjectGeneration
 {
