+++ Editors/Audio/BnkCompiler/ObjectGeneration/Warhammer3/ActionGenerator.cs
@@ -5,6 +5,7 @@ using Shared.GameFormats.Wwise.Hirc.V136;
 using Editors.Audio.BnkCompiler;
 using Editors.Audio.BnkCompiler.ObjectConfiguration.Warhammer3;
 using Editors.Audio.Utility;
+using Shared.GameFormats.Wwise.Hirc;
 
 namespace Editors.Audio.BnkCompiler.ObjectGeneration.Warhammer3
 {
