+++ Editors/Audio/BnkCompiler/ObjectGeneration/Warhammer3/EventGenerator.cs
@@ -6,6 +6,7 @@ using Shared.GameFormats.Wwise.Hirc.V136;
 using Editors.Audio.BnkCompiler;
 using Editors.Audio.BnkCompiler.ObjectConfiguration.Warhammer3;
 using Editors.Audio.BnkCompiler.ObjectGeneration;
+using Shared.GameFormats.Wwise.Hirc;
 
 namespace Editors.Audio.BnkCompiler.ObjectGeneration.Warhammer3
 {
