+++ Editors/Audio/BnkCompiler/ObjectGeneration/Warhammer3/RandomContainerGenerator.cs
@@ -5,6 +5,7 @@ using Shared.GameFormats.Wwise;
 using Shared.GameFormats.Wwise.Hirc.V136;
 using System.Collections.Generic;
 using Editors.Audio.BnkCompiler.ObjectConfiguration.Warhammer3;
+using Shared.GameFormats.Wwise.Hirc;
 
 namespace Editors.Audio.BnkCompiler.ObjectGeneration.Warhammer3
 {
