+++ Editors/Audio/BnkCompiler/ObjectGeneration/Warhammer3/SoundGenerator.cs
@@ -8,6 +8,7 @@ using Shared.GameFormats.Wwise;
 using Shared.GameFormats.Wwise.Hirc.V136;
 using Editors.Audio.BnkCompiler;
 using Editors.Audio.BnkCompiler.ObjectConfiguration.Warhammer3;
+using Shared.GameFormats.Wwise.Hirc;
 
 namespace Editors.Audio.BnkCompiler.ObjectGeneration.Warhammer3
 {
