+++ Editors/Reports/Audio/DialogueEventInfoPrinter.cs
@@ -2,7 +2,6 @@
 using Editors.Audio.Utility;
 using Shared.Core.Events;
 using Shared.Core.Misc;
-using Shared.GameFormats.Wwise;
 using Shared.GameFormats.Wwise.Hirc;
 
 namespace Editors.Reports.Audio
