+++ Shared/GameFiles/Wwise/Hirc/HircFactory.cs
@@ -22,6 +22,10 @@
             return new CAkUnknown();
         }
 
+        // Each major release of Wwise has a bank generator version.
+        // CA sometimes use in-house compiled version of Wwise which is based on a public release with custom modifications to some Wwise objects.
+        // The bank generator version of the closest public release (2019.2.15.7667) to that used in Wh3 (2147483784) is 135.
+        // Wwiser adds 1 to that for internal use to create a pseudo version called 136 but really it's 2147483784.
         public static HircFactory CreateFactory(uint version)
         {
             switch (version)
